Fix routing through new contorllers
